LayoutTests/http/tests/webgl/1.0.2/readPixelsBadArgs-expected.txt

CONSOLE MESSAGE: Promoted URL from http://www.opengl.org/img/opengl_logo.jpg to https
Blocked access to external URL https://www.opengl.org/img/opengl_logo.jpg
CONSOLE MESSAGE: Promoted URL from http://www.opengl.org/img/opengl_logo.jpg to https
Blocked access to external URL https://www.opengl.org/img/opengl_logo.jpg
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: resources/webgl_test_files/conformance/more/functions/readPixelsBadArgs.html
FAIL
